+++ tests/data_utils.py
@@ -1,7 +1,8 @@
 import tempfile
 import os
 import shutil
-from morgana.MLModel.io import new_model
+from morgana.MLModel.io import new_model, save_model
+from morgana.cli.train import train_model
 
 
 def create_test_model_folder(model, from_pretrained=False):
@@ -10,8 +11,7 @@ def create_test_model_folder(model, from_pretrained=False):
     model_folder = os.path.join(root, "model")
     model_tiny = "./tests/test_data/model_tiny"
     if from_pretrained:
-        os.mkdir(model_folder)
-        os.copytree(model_tiny, model_folder)
+        shutil.copytree(model_tiny, model_folder)
         pretrained = f"./tests/test_data/pretrained/{model}"
         [shutil.copy(os.path.join(pretrained, f), model_folder) for f in os.listdir(pretrained)]
     else:
@@ -25,3 +25,14 @@ def create_test_model_folder(model, from_pretrained=False):
             for f in os.listdir(os.path.join(model_tiny, "testset"))
         ]
     return model_folder
+
+
+def create_pretrained():
+    for model in ["logistic", "unet", "MLP"]:
+        model_folder = create_test_model_folder(model, from_pretrained=False)
+        classifier, scaler, params = train_model(model, model_folder=model_folder, epochs=10, steps_per_epoch=2)
+        save_model(f"./tests/test_data/pretrained/{model}", classifier, scaler, **params)
+
+
+if __name__ == "__main__":
+    create_pretrained()
