+++ tests/test_data/pretrained/logistic/params.json
@@ -1,0 +1,13 @@
+{
+    "bias": 0.4,
+    "down_shape": 0.25,
+    "edge_size": 5,
+    "feature_mode": "daisy",
+    "fraction": 0.25,
+    "model": "logistic",
+    "sigmas": [
+        1.0,
+        5.0,
+        15.0
+    ]
+}
